+++ contracts/Nuja/LeafNuja.sol
@@ -7,17 +7,17 @@ contract LeafNuja is Nuja {
         return '/ipfs/QmYzq4dHKY5JAEH97ABjTJui8aVUsQmzWg525YCzDovHrY';
     }
 
-    function power(uint8 x, uint8 y, uint8 player, uint[176] moveInput) public view returns(uint[176] moveOutput) {
+    function power(uint8 x, uint8 y, uint8 player, uint8[176] moveInput) public view returns(uint8[176] moveOutput) {
         var (r_x, r_y) = getPosition(moveInput, player);
 
         // Distance requirement
         require(distance(x, y, r_x, r_y) == 1);
 
         // Strike
-        uint opponent = getPlayer(moveInput, x, y);
+        uint8 opponent = getPlayer(moveInput, x, y);
         require(opponent > 0);
 
-        uint[176] memory tmp = damage(moveInput, opponent-1, 20);
+        uint8[176] memory tmp = damage(moveInput, opponent-1, 20);
         return restore(tmp, player, 10);
     }
 }
