+++ contracts/Weapon/Grenade.sol
@@ -9,18 +9,18 @@ contract Grenade is Weapon {
         return '/ipfs/QmWUHrPsJNB7wZNPHZKos4B1jqUMdwgvhDPf2ZPrrkuGof';
     }
 
-    function use(uint8 x, uint8 y, uint8 player, uint[176] moveInput) public view returns(uint[176] moveOutput) {
+    function use(uint8 x, uint8 y, uint8 player, uint8[176] moveInput) public view returns(uint8[176] moveOutput) {
         require(x > 0 && x < 7 && y > 0 && y < 7);
 
         var (r_x, r_y) = getPosition(moveInput, player);
 
         // Distance requirement
-        uint d = distance(x, y, r_x, r_y);
+        uint8 d = distance(x, y, r_x, r_y);
         require(d > 0 && d < 4);
 
-        uint[176] memory tmp = moveInput;
+        uint8[176] memory tmp = moveInput;
 
-        uint opponent = getPlayer(tmp, x-1, y-1);
+        uint8 opponent = getPlayer(tmp, x-1, y-1);
         if (opponent > 0) {
             tmp = damage(tmp, opponent-1, 30);
         }
