+++ contracts/Weapon/Sniper.sol
@@ -8,7 +8,7 @@ contract Sniper is Weapon {
         return '/ipfs/QmdPavz3brQLWyuNd97v8anNtggm95CBn63K5GA57RkpiK';
     }
 
-    function use(uint8 x, uint8 y, uint8 player, uint[176] moveInput) public view returns(uint[176] moveOutput) {
+    function use(uint8 x, uint8 y, uint8 player, uint8[176] moveInput) public view returns(uint8[176] moveOutput) {
         // We must be in a building to shoot
         var (r_x, r_y) = getPosition(moveInput, player);
         require(distance(x, y, r_x, r_y) > 2);
@@ -18,7 +18,7 @@ contract Sniper is Weapon {
         require(getBuilding(moveInput, x, y) == 0);
 
         // Strike
-        uint opponent = getPlayer(moveInput, x, y);
+        uint8 opponent = getPlayer(moveInput, x, y);
 
         require(opponent > 0);
         return damage(moveInput, opponent-1, 80);
