+++ contracts/Weapon/Weapon.sol
@@ -9,5 +9,5 @@ contract Weapon is Geometry, StateManager {
     function getMetadata() public pure returns (string metadata);
 
     // Function called by server to use the Weapon
-    function use(uint8 x, uint8 y, uint8 player, uint[176] moveInput) public view returns(uint[176] moveOutput);
+    function use(uint8 x, uint8 y, uint8 player, uint8[176] moveInput) public view returns(uint8[176] moveOutput);
 }
